Correction of forgotten initializations or missing type conversions, which makes code safer & a lot of justified compiler warnings disappear

+++ src/nnue/evaluate_nnue.cpp
@@ -109,7 +109,7 @@ namespace Stockfish::Eval::NNUE {
   {
     write_little_endian<std::uint32_t>(stream, Version);
     write_little_endian<std::uint32_t>(stream, hashValue);
-    write_little_endian<std::uint32_t>(stream, desc.size());
+    write_little_endian<std::uint32_t>(stream, static_cast<uint32_t>(desc.size()));
     stream.write(&desc[0], desc.size());
     return !stream.fail();
   }
